requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
aiomysql
python-jose[cryptography]
passlib[bcrypt]
pydantic
openai
python-dotenv
